beautify the code

+++ src/redis_hash.h
@@ -53,7 +53,7 @@ class Hash : public SubKeyScanner {
   rocksdb::Status IncrByFloat(const Slice &user_key, const Slice &field, double increment, double *ret);
   rocksdb::Status MSet(const Slice &user_key, const std::vector<FieldValue> &field_values, bool nx, int *ret);
   rocksdb::Status Range(const Slice &user_key, const Slice &start, const Slice& stop,
-                        int limit, std::vector<FieldValue> *field_values);
+                        int64_t limit, std::vector<FieldValue> *field_values);
   rocksdb::Status MGet(const Slice &user_key,
                        const std::vector<Slice> &fields,
                        std::vector<std::string> *values,
